pyproject.toml
[project]
name = "antex-backend"
version = "0.1.0"
description = "AntEx - Telegram Exchange Bot (RUB/USDT ↔ THB)"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.24.0",
    "aiohttp-socks>=0.10.1",
    "aiofiles>=25.1.0",
    "alembic>=1.18.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "fluent.runtime>=0.4.0",
    "httpx>=0.28.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.0",
    "python-dotenv>=1.1.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "uvloop>=0.22.0",
    "cryptography>=46.0.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "aiosqlite>=0.22.0",
    "ruff>=0.14.0",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
